(MAINT) This removes the uncessary v0 module declaration
Why In the go.mod the module was defined as being .v0.  It is not
neccessary to version the default (v0) version of a module.  It is implictlity a v0.
The format was also wrong as it should be a /vX as opposed to .vX.

How Remove the .v0 from the go.mod file and updated all deps.