(MAINT) If an operator is nil, it can cause a panic
What: Adding check to ensure we only use operator if it is non-nul.

Why: Its possible for a clause to have a nil operator, but the only valid
scenario for this, is if it is a segmentMatch operation.  We need protective code
to ensure we don't accidently call a nil operator.