[FFM-1355]: SDK doesnt support multiple custom attributes and name not set correctly
The Target builder replaces the attributes map with every new custom attribute.  This means
we only ever could have one custom attribute.

Name was stored as a custom attribute instead of as a named field, the problem is if you have
the Java SDK using the same target, then it sends 'name' as a named field, and so conflicts with
custom attributes.

This PR fixes the issue with the custom attributes and ensures name is stored in a named field.