[FFM-1452]: Fix bool, int and number comparisons + implement slices
This change fixes the bool, int and number comparisons.  It replaces
a cast with the appropriate function from the strconv package.

The slice type is implemented with support for in, contains and equals
operators.

Customers can store target attributes as strings, float64, bool and slices.
When creating a rule the value is always a slice of strings.

The code was trying to cast a string to a int, float etc which does not work.
This caused rules evaluations to fail.

Added unit tests to verify the type operators now work as expected.