FFM-1821 - Prerequisite rules not working on ff-server
There was a bug in the FF-Server where it wasn't evaluating flag values
properly if they had prerequisites. This is becuase the current logic
flow in FeatureConfig.Evaluate doesn't take Prerequisites into account
which meant client SDKs receiving evaluations for flags with
Prerequisites weren't receiving the correct value.

Since in order to evaluate prerequisites we need the flag we're
evaluating and all of the flags that may be prerequisites of it the only
way I could see to do this was to provide a way to pass in all other
flags. The way the Evaluator does it is by looking up the Prerequisite
flags from the cache but we don't have the option to do that from inside
the FeatureConfig type. So the next best thing I could think of was to
give the caller a way to pass in a map of flags that we could then do
the lookup against.