FFM-8067 Call onStreamError anytime we exit Subscribe
**What**

- Makes sure we call `onStreamError` anytime we break out of the
  SubscribeWithContext function
- If an eventListener has been passed to the Proxy we call its `Pub`
  method with an error. This lets the eventListener know that the stream
its listening to has disconnected

**Why**

- Currently we only call `onStreamError` if SubscribeWithContext returns
  an error. However if the stream is intentionally closed by the server
rather than bubbling up the EOF error the SSE library we're using
swallows it. This means we never call `onStreamError` which means the
SDK thinks it's still connected to the stream and receiving SSE events
which means it stops getting any changes because it isn't operating in
polling mode and won't get any SSE events. Calling `onStreamError` means we'll set `streamConnected=false` and the
SDK will poll for changes until it can reconnect to the stream
- For the eventListner code we can use this in the ff-proxy to pass on
  disconnects between the Proxy & ff-server to SDKs connected to the
Proxy

**Testing**

I've been testing this out locally by getting the Proxy to close the
stream with the SDK if it disconnects from ff-server