FFM-8618 Add SetFlags & SetSegments methods
**What**

- Adds a SetFlags and SetSegment methods to the repository

**Why**

The Proxy passes its cache to the SDK and lets it manage populating and
updating the cache with flags & segments. The current SetFlag and
SetSegment methods work for storing individual key/value pairs of flags
and segments. But we need to be able to store all of the flags/segments
for a given environment at a single key.

**Testing**

I updated my local proxy to use this version of the sdk and did some
manual testing and its populating & updating the cache properly when I
start the proxy up and then toggle/edit a flag/segment