[FFM-9476]: Optimize memory usage (#126)
This change supports improved memory usage of the evaluator code.
It introduces a new function on the query interface GetFlagMap().  This returns a map of flags.

This is more optimal as the GetFlag function can also use a underlying map, it removes the need to maintain both a list
and a map of flags for the caller of the evaluator.